Rename vacancy handler
